requirements.txt
scipy==1.5.4
torch==1.6.0+cpu
pyqtgraph==0.11.1
numpy==1.19.5
mayavi==4.7.4
tqdm==4.42.1
pandas==1.1.5
opencv_python_headless==4.5.3.56
torchvision==0.7.0+cpu
matplotlib==3.1.2
Pillow==9.0.0
PyQt5==5.15.6
